Add rudimentary support for CPU detection
The file /proc/cpuinfo is parsed to get information about the CPU. Use entry "model name" for CPU model and the last entry for "processor" + 1 for thread count.
On my machine, the model name includes unnecessary spaces, so more than one consecutive space is removed from the model name. Also the CPU frequency is included in the model name. This is used for now, but should be changed to the actual frequency.
Tested with Intel Core 2 Duo T7200 and Intel Core i5-2540m