Get name of terminal emulator from X server
This assumes that the class of the active window matches the name of the
terminal emulator being used to run paleofetch, which seems a safe bet.

We throw away a lot of the result values from `XGetWindowProperty()` on
the belief that the requests are exceedingly unlikely to fail, which is
especially true if we're taking as a given that an X server is running.