Fix XopenDisplay error when running without Xorg in tty
Add checks if xserver is running and replace information that is gathered from xserver with alternatives.
This affects resolution and terminal name. Terminal name is replaced with the old $TERM method, when running in tty.
For resolution, the available display modes in /sys/class/drm are parsed and the highest avaiblable resolution is displayed.
Works on my machine, but should be tested on other computers.