Make substring removing for CPU and GPU configurable
This adds a configurable way to remove substrings from CPU and GPU model names.
I was not able to make it work with the new REPLACE_CONST_STRING macro. It would be nice if you could omit the substring lenght in the config, maybe somebody has a neat idea for this.